Make some Signals validate and helpers fn generic.
