+++ public/apps/signals/utils/helpers.js
@@ -6,17 +6,12 @@ export function stringifyPretty(obj) {
 
 export const isNameHasProhibitedSpecialChars = (name = '') => (/[^\w\s_-]+/gm).test(name);
 
-export const arrayToComboBoxOptions = (array = []) => {
-  if (!array) return [];
-  return array.map(label => ({ label })).sort((a, b) => a.label.localeCompare(b.label));
-};
-
-export const comboBoxOptionsToArray = (array = []) => {
-  if (!array) return [];
-  return array.map(({ label }) => label).sort((a, b) => a.localeCompare(b));
-};
-
 export const matchTimeHHMM = timeStr => timeStr.match(/^(0[0-9]|1[0-9]|2[0-3]|[0-9]):[0-5][0-9]$/);
 export const matchDayOfMonth = dayOfMonthStr => (dayOfMonthStr + '').match(/^([1-9]|[12]\d|3[01])$/);
 export const matchTimePeriod = timePeriodStr => timePeriodStr.match(/(\d+)([YMwdhms])$/);
 export const matchDayOfWeek = dayOfWeekStr => dayOfWeekStr.match(/\b((mon|tues|wed(nes)?|thu(rs)?|fri|sat(ur)?|sun)(day)?)\b/);
+
+export {
+  arrayToComboBoxOptions,
+  comboBoxOptionsToArray
+} from '../../utils/helpers';
