+++ public/apps/signals/utils/i18n/common.js
@@ -1,68 +1,1 @@
-import React from 'react';
-import { EuiI18n } from '@elastic/eui';
-
-export const helpText = (<EuiI18n token="sg.common.help.text" default="Help" />);
-export const nameText = (<EuiI18n token="sg.common.name.text" default="Name" />);
-export const generalText = (<EuiI18n token="sg.common.general.text" default="General" />);
-export const cancelText = (<EuiI18n token="sg.common.cancel.text" default="Cancel" />);
-export const confirmText = (<EuiI18n token="sg.common.confirm.text" default="Confirm" />);
-export const saveText = (<EuiI18n token="sg.common.save.text" default="Save" />);
-export const deleteText = (<EuiI18n token="sg.common.delete.text" default="Delete" />);
-export const updateText = (<EuiI18n token="sg.common.update.text" default="Update" />);
-export const createText = (<EuiI18n token="sg.common.create.text" default="Create" />);
-export const editText = (<EuiI18n token="sg.common.edit.text" default="Edit" />);
-export const definitionText = (<EuiI18n token="sg.common.definition.text" default="Definition" />);
-export const graphText = (<EuiI18n token="sg.common.graph.text" default="Graph" />);
-export const jsonText = (<EuiI18n token="sg.common.json.text" default="JSON" />);
-export const typeText = (<EuiI18n token="sg.common.type.text" default="Type" />);
-export const responseText = (<EuiI18n token="sg.common.reponse.text" default="Response" />);
-export const actionText = (<EuiI18n token="sg.common.action.text" default="Action" />);
-export const actionsText = (<EuiI18n token="sg.common.actions.text" default="Actions" />);
-export const executeText = (<EuiI18n token="sg.common.execute.text" default="Execute" />);
-export const activeText = (<EuiI18n token="sg.common.active.text" default="Active" />);
-export const urlText = (<EuiI18n token="sg.common.url.text" default="Url" />);
-export const bodyText = (<EuiI18n token="sg.common.body.text" default="Body" />);
-export const headersText = (<EuiI18n token="sg.common.headers.text" default="Headers" />);
-export const methodText = (<EuiI18n token="sg.common.method.text" default="Method" />);
-export const throttlePeriodText = (<EuiI18n token="sg.common.throttlePeriod.text" default="Throttle Period" />);
-export const noText = (<EuiI18n token="sg.common.no.text" default="No" />);
-export const yesText = (<EuiI18n token="sg.common.yes.text" default="Yes" />);
-export const cloneText = (<EuiI18n token="sg.common.clone.text" default="Clone" />);
-export const inspectText = (<EuiI18n token="sg.common.inspect.text" default="Inspect" />);
-export const requiredText = (<EuiI18n token="sg.common.required.text" default="Required" />);
-export const confirmDeleteText = (<EuiI18n token="sg.common.confirmDelete.text" default="Confirm Delete" />);
-export const problemWithValidationTryAgainText = (
-  <EuiI18n token="sg.common.problemWithValidationTryAgain.text" default="Problem with validation. Please try again" />
-);
-export const nameAlreadyExistsText = (
-  <EuiI18n token="sg.common.nameAlreadyExists.text" default="Name already exists" />
-);
-export const mustBePositiveIntegerText = (
-  <EuiI18n token="sg.common.mustBePositiveInteger.text" default="Must be positive integer" />
-);
-export const mustBeNumberBetween1And31Text = (
-  <EuiI18n token="sg.common.mustBeNumberBetween1And31.text" default="Must be number between 1 and 31" />
-);
-export const mustSpecifyIndexText = (
-  <EuiI18n token="sg.common.mustSpecifyIndex.text" default="Must specify index" />
-);
-export const mustSpecifyValueText = (
-  <EuiI18n token="sg.common.mustSpecifyValue.text" default="Must specify value" />
-);
-export const doYouReallyWantToDeleteText = (
-  <EuiI18n token="sg.common.doYouReallyWantToDelete.text" default="Do you really want to delete" />
-);
-export const documentationText = (
-  <EuiI18n token="sg.common.documentation.text" default="Documentation" />
-);
-export const validCharsForNameAreText = (
-  <EuiI18n token="sg.common.validCharsForNameAre.text" default="Valid characters for name are letter, number, space, underscore and dash" />
-);
-export const addText = (<EuiI18n token="sg.common.addText.text" default="Add" />);
-export const closeText = (<EuiI18n token="sg.common.close.text" default="Close" />);
-export const invalidJsonText = (<EuiI18n token="sg.common.invalidJson.text" default="Invalid JSON" />);
-export const securityText = (<EuiI18n token="sg.common.security.text" default="Security" />);
-export const userText = (<EuiI18n token="sg.common.user.text" default="User" />);
-export const passwordText = (<EuiI18n token="sg.common.password.text" default="Password" />);
-export const defaultsText = (<EuiI18n token="sg.common.defaults.text" default="Defaults" />);
-export const invalidEmailAddressText = (<EuiI18n token="sg.common.invalidEmailAddress.text" default="Invalid email address" />);
+export * from '../../../utils/i18n/common';
