+++ public/apps/signals/utils/validate/validateName.js
