+++ public/apps/utils/helpers/arrayToComboBoxOptions.js
@@ -1,0 +1,4 @@
+export const arrayToComboBoxOptions = (array = []) => {
+  if (!array) return [];
+  return array.map(label => ({ label })).sort((a, b) => a.label.localeCompare(b.label));
+};
