+++ public/apps/utils/helpers/comboBoxOptionsToArray.js
@@ -1,0 +1,4 @@
+export const comboBoxOptionsToArray = (array = []) => {
+  if (!array) return [];
+  return array.map(({ label }) => label).sort((a, b) => a.localeCompare(b));
+};
