+++ public/apps/utils/i18n/common.js
@@ -8,23 +8,9 @@ export const cancelText = (<EuiI18n token="sg.common.cancel.text" default="Cance
 export const confirmText = (<EuiI18n token="sg.common.confirm.text" default="Confirm" />);
 export const saveText = (<EuiI18n token="sg.common.save.text" default="Save" />);
 export const deleteText = (<EuiI18n token="sg.common.delete.text" default="Delete" />);
-export const updateText = (<EuiI18n token="sg.common.update.text" default="Update" />);
-export const createText = (<EuiI18n token="sg.common.create.text" default="Create" />);
-export const editText = (<EuiI18n token="sg.common.edit.text" default="Edit" />);
-export const definitionText = (<EuiI18n token="sg.common.definition.text" default="Definition" />);
-export const graphText = (<EuiI18n token="sg.common.graph.text" default="Graph" />);
-export const jsonText = (<EuiI18n token="sg.common.json.text" default="JSON" />);
-export const typeText = (<EuiI18n token="sg.common.type.text" default="Type" />);
-export const responseText = (<EuiI18n token="sg.common.reponse.text" default="Response" />);
-export const actionText = (<EuiI18n token="sg.common.action.text" default="Action" />);
-export const actionsText = (<EuiI18n token="sg.common.actions.text" default="Actions" />);
-export const executeText = (<EuiI18n token="sg.common.execute.text" default="Execute" />);
-export const activeText = (<EuiI18n token="sg.common.active.text" default="Active" />);
 export const urlText = (<EuiI18n token="sg.common.url.text" default="Url" />);
 export const bodyText = (<EuiI18n token="sg.common.body.text" default="Body" />);
 export const headersText = (<EuiI18n token="sg.common.headers.text" default="Headers" />);
-export const methodText = (<EuiI18n token="sg.common.method.text" default="Method" />);
-export const throttlePeriodText = (<EuiI18n token="sg.common.throttlePeriod.text" default="Throttle Period" />);
 export const noText = (<EuiI18n token="sg.common.no.text" default="No" />);
 export const yesText = (<EuiI18n token="sg.common.yes.text" default="Yes" />);
 export const cloneText = (<EuiI18n token="sg.common.clone.text" default="Clone" />);
@@ -63,3 +49,22 @@ export const closeText = (<EuiI18n token="sg.common.close.text" default="Close" 
 export const invalidJsonText = (<EuiI18n token="sg.common.invalidJson.text" default="Invalid JSON" />);
 export const errorText = (<EuiI18n token="sg.common.error.text" default="Error" />);
 export const successText = (<EuiI18n token="sg.common.success.text" default="Success" />);
+export const invalidEmailAddressText = (<EuiI18n token="sg.common.invalidEmailAddress.text" default="Invalid email address" />);
+export const updateText = (<EuiI18n token="sg.common.update.text" default="Update" />);
+export const createText = (<EuiI18n token="sg.common.create.text" default="Create" />);
+export const editText = (<EuiI18n token="sg.common.edit.text" default="Edit" />);
+export const definitionText = (<EuiI18n token="sg.common.definition.text" default="Definition" />);
+export const graphText = (<EuiI18n token="sg.common.graph.text" default="Graph" />);
+export const jsonText = (<EuiI18n token="sg.common.json.text" default="JSON" />);
+export const typeText = (<EuiI18n token="sg.common.type.text" default="Type" />);
+export const responseText = (<EuiI18n token="sg.common.reponse.text" default="Response" />);
+export const actionText = (<EuiI18n token="sg.common.action.text" default="Action" />);
+export const actionsText = (<EuiI18n token="sg.common.actions.text" default="Actions" />);
+export const executeText = (<EuiI18n token="sg.common.execute.text" default="Execute" />);
+export const activeText = (<EuiI18n token="sg.common.active.text" default="Active" />);
+export const methodText = (<EuiI18n token="sg.common.method.text" default="Method" />);
+export const throttlePeriodText = (<EuiI18n token="sg.common.throttlePeriod.text" default="Throttle Period" />);
+export const securityText = (<EuiI18n token="sg.common.security.text" default="Security" />);
+export const userText = (<EuiI18n token="sg.common.user.text" default="User" />);
+export const passwordText = (<EuiI18n token="sg.common.password.text" default="Password" />);
+export const defaultsText = (<EuiI18n token="sg.common.defaults.text" default="Defaults" />);
