+++ public/apps/utils/validate/hasError.js
