+++ public/apps/utils/validate/validateEmptyArray.js
