+++ public/apps/utils/validate/validateEmptyField.js
