+++ public/apps/utils/validate/validateIndex.js
