+++ public/apps/utils/validate/validateInterval.js
